+++ src/main/java/net/caffeinemc/sodium/render/chunk/state/ChunkRenderBounds.java
@@ -3,13 +3,15 @@ package net.caffeinemc.sodium.render.chunk.state;
 import net.minecraft.util.math.ChunkSectionPos;
 
 public class ChunkRenderBounds {
-    public static final ChunkRenderBounds ALWAYS_FALSE = new ChunkRenderBounds(Float.POSITIVE_INFINITY, Float.POSITIVE_INFINITY, Float.POSITIVE_INFINITY,
-            Float.NEGATIVE_INFINITY, Float.NEGATIVE_INFINITY, Float.NEGATIVE_INFINITY);
+    public static final ChunkRenderBounds ALWAYS_FALSE =new ChunkRenderBounds(
+            Double.POSITIVE_INFINITY, Double.POSITIVE_INFINITY, Double.POSITIVE_INFINITY,
+            Double.NEGATIVE_INFINITY, Double.NEGATIVE_INFINITY, Double.NEGATIVE_INFINITY
+    );
 
-    public final float x1, y1, z1;
-    public final float x2, y2, z2;
+    public final double x1, y1, z1;
+    public final double x2, y2, z2;
 
-    public ChunkRenderBounds(float x1, float y1, float z1, float x2, float y2, float z2) {
+    public ChunkRenderBounds(double x1, double y1, double z1, double x2, double y2, double z2) {
         this.x1 = x1;
         this.y1 = y1;
         this.z1 = z1;
@@ -58,13 +60,13 @@ public class ChunkRenderBounds {
             int z2 = origin.getMinZ() + rightBound(this.z);
 
             return new ChunkRenderBounds(
-                    Math.max(x1, origin.getMinX()) - 0.5f,
-                    Math.max(y1, origin.getMinY()) - 0.5f,
-                    Math.max(z1, origin.getMinZ()) - 0.5f,
+                    Math.max(x1, origin.getMinX()) - 0.5d,
+                    Math.max(y1, origin.getMinY()) - 0.5d,
+                    Math.max(z1, origin.getMinZ()) - 0.5d,
 
-                    Math.min(x2, origin.getMaxX()) + 0.5f,
-                    Math.min(y2, origin.getMaxY()) + 0.5f,
-                    Math.min(z2, origin.getMaxZ()) + 0.5f
+                    Math.min(x2, origin.getMaxX()) + 0.5d,
+                    Math.min(y2, origin.getMaxY()) + 0.5d,
+                    Math.min(z2, origin.getMaxZ()) + 0.5d
             );
         }
